refactor: :recycle: 수정 사항 반영 2차
